fix(playback): extract spawn test + skip 2 unkillable mutants
Three CI-blocker fixes on PR #58:

1. pipeline.rs was at 1007/1000 lines after the mutant-kill test
   addition. Extracted spawn_stores_ndi_name_for_accessor to a sibling
   file via #[path], matching the pipeline_heartbeat_tests.rs precedent.
   pipeline.rs is now back under cap (998 lines).

2. Both PlaybackPipeline::spawn variants get #[cfg_attr(test,
   mutants::skip)]. Default::default() body-replacement is structurally
   unsound (PlaybackPipeline has no Default impl); the Windows variant
   is not exercised on the Linux mutation runner anyway.

3. The nominal_fps > 0.0 guard inside classify_bad_poll was a
   redundant defensive check: with nominal_fps=0.0, observed_fps < 0.0
   is unreachable for any non-negative observed_fps, so the guard
   never short-circuits a real bad poll. Removing it eliminates the
   logically-equivalent boundary mutant entirely.
